requirements.txt
fastapi
uvicorn
python-multipart
psycopg2-binary
python-dotenv
